# fixes #34: improve performance by avoiding flushing the StringBuffer each row