It turns out indeed that there was a performance issue. Namely, for each call to `convertSingleRow`, `sb.toString()` is called, which builds up a `String` of the entire CSV-output so far. The size of the output increases with each row, and therefor makes the conversion slower and slower. The return value was actually not used by `convert`.

To keep backwards compatibility with users that rely on the returned strings, I've added an option `returnString` which is `true` by default. The `convert` method sets it to `false`, so that it is optimized.

Users running into performance problems should pass `returnString: false` when calling `convertSingleRow` manually, and only use the `StringBuffer sb`.