@boukeversteegh 2020-10-31
I forgot the mention the magnitude of the performance increase.

| Number of rows | Before | After |
|----------:|------------:|------:|
|100|14ms|11ms|
|1000|46ms|32ms|
|10000|1.9s|117ms|
|50000|1m23s|398ms|
|100000|?|756ms|

I needed to output a CSV with 500k records. It would most likely take an infinite amount of time to produce it. Now it takes about a second.