@close2 2020-10-31
Thanks.
`convertSingleRow` shouldn't call `toString`.

I guess this is a relict from a previous version.

Having this option is at least a good work around.
The next major version will remove the return `String`.