@boukeversteegh 2020-11-01
Yes, I thought so 😃 Thanks for merging and releasing a new version so quickly!